+++ esp-hal/src/gpio/rtc_io.rs
@@ -33,7 +33,7 @@ use core::marker::PhantomData;
 
 use super::{InputPin, OutputPin, RtcPin};
 use crate::{
-    gpio::RtcFunction,
+    gpio::{RtcFunction, WakeEvent},
     peripherals::{GPIO, RTC_IO},
 };
 
@@ -92,6 +92,7 @@ impl<'d, const PIN: u8> LowPowerInput<'d, PIN> {
         this.input_enable(true);
         this.pullup_enable(false);
         this.pulldown_enable(false);
+        this.wakeup_enable(None);
 
         this
     }
@@ -114,6 +115,17 @@ impl<'d, const PIN: u8> LowPowerInput<'d, PIN> {
         RTC_IO::regs()
             .touch_pad(PIN as usize)
             .modify(|_, w| w.rde().bit(enable));
+    }
+
+    /// Allows this pin to wakeup the ULP core, when UlpCoreWakeupSource::Gpio is used.
+    pub fn wakeup_enable(&self, event: Option<WakeEvent>) {
+        let pin_reg = GPIO::regs().pin(PIN as usize);
+
+        if let Some(evt) = event {
+            pin_reg.write(|w| unsafe { w.int_type().bits(evt as u8).wakeup_enable().bit(true) });
+        } else {
+            pin_reg.write(|w| w.wakeup_enable().bit(false));
+        }
     }
 }
 
@@ -139,6 +151,7 @@ impl<'d, const PIN: u8> LowPowerOutputOpenDrain<'d, PIN> {
         this.pullup_enable(true);
         this.pulldown_enable(false);
         this.output_enable(true);
+        this.wakeup_enable(None);
 
         this
     }
@@ -181,5 +194,16 @@ impl<'d, const PIN: u8> LowPowerOutputOpenDrain<'d, PIN> {
         GPIO::regs()
             .pin(PIN as usize)
             .modify(|_, w| w.pad_driver().bit(enable));
+    }
+
+    /// Allows this pin to wakeup the ULP core, when UlpCoreWakeupSource::Gpio is used.
+    pub fn wakeup_enable(&self, event: Option<WakeEvent>) {
+        let pin_reg = GPIO::regs().pin(PIN as usize);
+
+        if let Some(evt) = event {
+            pin_reg.write(|w| unsafe { w.int_type().bits(evt as u8).wakeup_enable().bit(true) });
+        } else {
+            pin_reg.write(|w| w.wakeup_enable().bit(false));
+        }
     }
 }