+++ esp-lp-hal/src/gpio/interrupt.rs
@@ -62,16 +62,10 @@ pub fn gpio_interrupt_clear(pinmask: u32) {
 fn gpio_interrupt_disable(pin_mask: u32) {
     // expects pinmask bit 0 == GPIO0
     let pin_iter = InterruptStatus::from(pin_mask).iterator();
-    let gpio_bank_reg = LpIo::ptr() as usize;
-    let int_type_mask: u32 = 0xFFFFFFFF ^ (0b111 << 7);
-
     for n in pin_iter {
-        let gpio_base = gpio_bank_reg + 0x28 + ((n as usize) * 4);
-        let gpio_pin = gpio_base as *mut u32;
-        // Read settings, clear int_type flags, write settings
-        let mut pin_setting = unsafe { gpio_pin.read_volatile() };
-        pin_setting &= int_type_mask;
-        unsafe { gpio_pin.write_volatile(pin_setting) }
+        unsafe { &*LpIo::PTR }
+            .pin(n as usize)
+            .write(|w| unsafe { w.int_type().bits(0) });
     }
 }
 
@@ -81,39 +75,16 @@ fn gpio_interrupt_disable(pin_mask: u32) {
 /// - `int_type`: interrupt type code, value from [Event], 0 to disable interrupts. If None, will
 ///   leave the int_type setting as-is.
 /// - `wake_up`: whether to wake up from light sleep.
-pub fn enable_pin_interrupt<const N: u8>(int_type: Option<u8>, wakeup_enable: bool) {
-    // let GPIO_BASE = unsafe { &*LpIo::PTR }.pin0().as_ptr().addr();
-    // TODO: Add LpIo::regs().pins(number : usize) to esp-pacs
-    let gpio_bank_reg = LpIo::ptr() as usize;
-    let gpio_base = gpio_bank_reg + 0x28 + ((N as usize) * 4);
-    let gpio_pin = gpio_base as *mut u32;
+pub fn enable_pin_interrupt<const N: u8>(int_type: u8) {
+    let gpio_pin = unsafe { &*LpIo::PTR }.pin(N as usize);
 
-    // Read the current setting
-    let mut pin_setting = unsafe { gpio_pin.read_volatile() };
-
-    // Write int_type if specified
-    if let Some(int_type) = int_type {
-        // Bits 7:9
-        // - 0: GPIO interrupt disable
-        // - 1: rising edge trigger
-        // - 2: falling edge trigger
-        // - 3: any edge trigger
-        // - 4: low level trigger
-        // - 5: high level trigger
-        let int_type_mask: u32 = 0xFFFFFFFF ^ (0b111 << 7);
-        pin_setting &= int_type_mask;
-        pin_setting |= (int_type as u32) << 7;
-    }
-
-    // Bit 10, wakeup enable.
-    // Used with UlpCoreWakeupSource::Gpio.
-    let wakeup_en_mask: u32 = 0xFFFFFFFF ^ (0b1 << 10);
-    pin_setting &= wakeup_en_mask;
-    if wakeup_enable {
-        pin_setting |= 0b1 << 10;
-    }
-
-    unsafe { gpio_pin.write_volatile(pin_setting) }
+    // - 0: GPIO interrupt disable
+    // - 1: rising edge trigger
+    // - 2: falling edge trigger
+    // - 3: any edge trigger
+    // - 4: low level trigger
+    // - 5: high level trigger
+    gpio_pin.write(|w| unsafe { w.int_type().bits(int_type) });
 }
 
 /// Clear pin interrupt
@@ -129,6 +100,11 @@ pub fn is_interrupt_set<const N: u8>() -> bool {
 
 /// Global GPIO wakeup enable/disable
 pub fn gpio_wakeup_enable(enable: bool) {
+    #[cfg(esp32s2)]
+    unsafe { &*crate::pac::RTC_CNTL::PTR }
+        .ulp_cp_timer()
+        .write(|w| w.ulp_cp_gpio_wakeup_ena().bit(enable));
+    #[cfg(esp32s3)]
     unsafe { &*crate::pac::RTC_CNTL::PTR }
         .rtc_ulp_cp_timer()
         .write(|w| w.ulp_cp_gpio_wakeup_ena().bit(enable));
@@ -136,6 +112,11 @@ pub fn gpio_wakeup_enable(enable: bool) {
 
 /// Clear GPIO wakeup status
 pub fn gpio_wakeup_clear() {
+    #[cfg(esp32s2)]
+    unsafe { &*crate::pac::RTC_CNTL::PTR }
+        .ulp_cp_timer()
+        .write(|w| w.ulp_cp_gpio_wakeup_clr().set_bit());
+    #[cfg(esp32s3)]
     unsafe { &*crate::pac::RTC_CNTL::PTR }
         .rtc_ulp_cp_timer()
         .write(|w| w.ulp_cp_gpio_wakeup_clr().set_bit());