+++ examples/peripheral/lp_core/ulp_timer_counter/src/main.rs
@@ -10,9 +10,7 @@
 //! The following wiring is assumed:
 //! - BUTTON => GPIO0
 
-//% CHIPS: esp32s3
-//! (ESP32S2 also supported, but not listed above, as the interrupt_counter example is only built
-//! for esp32s3 in CI)
+//% CHIPS: esp32s3,esp32s2
 
 #![no_std]
 #![no_main]